refactor(emitter): split system_emit.rs to satisfy §19 hard file-size limit
system_emit.rs was 2093 lines, exceeding the 2000-line hard limit from
CLAUDE.md §19. Extract the five `using`-declaration emit helpers into a
new sibling module `system_emit_using.rs` (717 lines). The split is a
pure code-motion: no logic changes, no behavior changes.

Visibility adjustments made to preserve inter-module call paths:
- `emit_system_top_level_using_scope` → pub(super) (called from execute_body)
- `emit_system_import_equals_declaration` → pub(super) (called from using scope)
- `emit_system_export_declaration` → pub(super) (called from using scope)
- `emit_system_enum_with_export_fold` → pub(super) (called from using scope)
- `emit_system_namespace_with_export_fold` → pub(super) (called from using scope)

Result: system_emit.rs 2093 → 1383 lines, system_emit_using.rs 717 lines.
